more fixes
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@163507 13f79535-47bb-0310-9956-ffa450edef68

+++ sandbox/pom-pre-alpha-converter/src/main/java/org/apache/maven/tools/converter/Main.java
@@ -72,6 +72,11 @@ public class Main
 
             if ( reverse )
             {
+                if ( !backup.exists() )
+                {
+                    System.err.println( "skipping... backup exists" );
+                    continue;
+                }
                 FileUtils.copyFile( backup, file );
                 backup.delete();
             }
@@ -85,6 +90,7 @@ public class Main
                     if ( backup.exists() )
                     {
                         System.err.println( "skipping... backup exists" );
+                        continue;
                     }
 
                     FileReader fileReader = new FileReader( file );
